+++ src/Observers/AttributeOptionValueObserver.php
@@ -1,0 +1,151 @@
+<?php
+
+/**
+ * TechDivision\Import\Attribute\Observers\AttributeOptionValueObserver
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-attribute
+ * @link      http://www.techdivision.com
+ */
+
+namespace TechDivision\Import\Attribute\Observers;
+
+use TechDivision\Import\Utils\StoreViewCodes;
+use TechDivision\Import\Attribute\Utils\ColumnKeys;
+use TechDivision\Import\Attribute\Utils\MemberNames;
+use TechDivision\Import\Subjects\SubjectInterface;
+use TechDivision\Import\Attribute\Services\AttributeBunchProcessorInterface;
+
+/**
+ * Observer that create's the attribute option values found in the additional CSV file.
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-attribute
+ * @link      http://www.techdivision.com
+ */
+class AttributeOptionValueObserver extends AbstractAttributeImportObserver
+{
+
+    /**
+     * The attribute processor instance.
+     *
+     * @var \TechDivision\Import\Attribute\Services\AttributeBunchProcessorInterface
+     */
+    protected $attributeBunchProcessor;
+
+    /**
+     * Initializes the observer with the passed subject instance.
+     *
+     * @param \TechDivision\Import\Subjects\SubjectInterface                           $subject                 The observer's subject instance
+     * @param \TechDivision\Import\Attribute\Services\AttributeBunchProcessorInterface $attributeBunchProcessor The attribute bunch processor instance
+     */
+    public function __construct(
+        SubjectInterface $subject,
+        AttributeBunchProcessorInterface $attributeBunchProcessor
+    ) {
+
+        // pass the subject through to the parend observer
+        parent::__construct($subject);
+
+        // initialize the attribute bunch processor
+        $this->attributeBunchProcessor = $attributeBunchProcessor;
+    }
+
+    /**
+     * Process the observer's business logic.
+     *
+     * @return void
+     */
+    protected function process()
+    {
+
+        // prepare the store view code
+        $this->prepareStoreViewCode();
+
+        // prepare and insert the attribute option value
+        $this->persistAttributeOptionValue($this->initializeAttribute($this->prepareAttributes()));
+    }
+
+    /**
+     * Prepare the attributes of the entity that has to be persisted.
+     *
+     * @return array The prepared attributes
+     */
+    protected function prepareAttributes()
+    {
+
+        // load the option ID
+        $optionId = $this->getLastOptionId();
+
+        // load the store ID
+        $storeId = $this->getRowStoreId(StoreViewCodes::ADMIN);
+
+        // load the attribute option value
+        $value = $this->getValue(ColumnKeys::VALUE);
+
+        // return the prepared attribute option
+        return $this->initializeEntity(
+            array(
+                MemberNames::OPTION_ID  => $optionId,
+                MemberNames::STORE_ID   => $storeId,
+                MemberNames::VALUE      => $value
+            )
+        );
+    }
+
+    /**
+     * Initialize the attribute with the passed attributes and returns an instance.
+     *
+     * @param array $attr The attribute attributes
+     *
+     * @return array The initialized attribute
+     */
+    protected function initializeAttribute(array $attr)
+    {
+        return $attr;
+    }
+
+    /**
+     * Return's the attribute bunch processor instance.
+     *
+     * @return \TechDivision\Import\Attribute\Services\AttributeBunchProcessorInterface The attribute bunch processor instance
+     */
+    protected function getAttributeBunchProcessor()
+    {
+        return $this->attributeBunchProcessor;
+    }
+
+    /**
+     * Return's the ID of the option that has been created recently.
+     *
+     * @return integer The option ID
+     */
+    protected function getLastOptionId()
+    {
+        return $this->getSubject()->getLastOptionId();
+    }
+
+    /**
+     * Persist the passed attribute option value.
+     *
+     * @param array $attributeOptionValue The attribute option value to persist
+     *
+     * @return void
+     */
+    protected function persistAttributeOptionValue(array $attributeOptionValue)
+    {
+        return $this->getAttributeBunchProcessor()->persistAttributeOptionValue($attributeOptionValue);
+    }
+}