+++ CHANGELOG.md
@@ -1,12 +1,22 @@
+# Version 1.0.0-alpha3
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Implement functionality to persist EAV attribute options + labels (no update)
+
 # Version 1.0.0-alpha2
 
 ## Bugfixes
 
 * None
 
-## Implement functionality to persist EAV attribute without labels + options
+## Features
 
-* Initial Release
+* Implement functionality to persist EAV attribute without labels + options
 
 # Version 1.0.0-alpha1